Add files via upload
This ADF pipeline is built to dynamically and flexibly load raw data files into an Azure Blob Storage Container. It is parameterized, making it an essential pipeline for this process. The initial and the important step for end to end process. This is the template of pipeline exported from Azure Data Factory.